successfully enabled a delete feature on the job issues table
