+++ src/redux/sagas/_root.saga.js
@@ -17,7 +17,7 @@ import getCampaignsSaga from './getCampaign.saga';
 import patchTemplateSaga from './patchTemplate.saga';
 import getTemplatesSaga from './getTemplates.saga';
 import postNewTemplateSaga from './postTemplate.saga';
-import getJobIssuesSaga from './getJobIssue.saga';
+import jobIssuesSaga from './jobIssue.saga';
 
 
 // rootSaga is the primary saga.
@@ -47,6 +47,6 @@ export default function* rootSaga() {
     patchTemplateSaga(), //PUT..er...patch - fix up template for new send
     getTemplatesSaga(), //GET all templates (for dropdown when building email)
     postNewTemplateSaga(), //POST new template from scratch
-    getJobIssuesSaga(), //GET job issues from database
+    jobIssuesSaga(), //GET job issues from database
   ]);
 }
